+++ developer-extension/src/panel/components/tabs/settingsTab.tsx
@@ -2,7 +2,7 @@ import { Badge, Box, Checkbox, Code, Group, Select, Space, Text } from '@mantine
 import React from 'react'
 import { DevServerStatus, useDevServerStatus } from '../../hooks/useDevServerStatus'
 import { useSettings } from '../../hooks/useSettings'
-import type { EventSource } from '../../types'
+import type { EventSource } from '../../hooks/useEvents'
 import { Columns } from '../columns'
 import { TabBase } from '../tabBase'
