+++ developer-extension/src/panel/hooks/useEvents/index.ts
@@ -1,0 +1,3 @@
+export { useEvents } from './useEvents'
+export { EventFilters } from './eventFilters'
+export { EventSource, StoredEvent } from './eventCollection'
