--- developer-extension/src/panel/types.ts
@@ -1,1 +1,0 @@
-export type EventSource = 'sdk' | 'requests'
